*** empty log message ***
git-svn-id: file:///usr/local/svn/fsc2@3639 934aee4a-8510-0410-ae50-ec5b2ea4bb86

+++ trunk/fsc2/src/func_basic.c
@@ -1696,7 +1696,7 @@ Var *f_islice( Var *v )
 		THROW( EXCEPTION )
 	}
 
-	array = T_calloc( size, sizeof( long ) );
+	array = T_calloc( ( size_t ) size, sizeof( long ) );
 	ret = vars_push( INT_ARR, array, ( size_t ) size );
 	T_free( array );
